+++ examples/storage/blob/variables.tf
@@ -1,0 +1,34 @@
+variable region_name {
+  description = "The AWS region to deploy into (e.g. eu-central-1)."
+  type = string
+}
+
+variable solution_name {
+  description = "The name of the AWS solution that owns all AWS resources."
+  type = string
+}
+
+variable solution_stage {
+  description = "The name of the current AWS solution stage."
+  type = string
+}
+
+variable solution_fqn {
+  description = "The fully qualified name of the current AWS solution."
+  type = string
+}
+
+variable common_tags {
+  description = "Common tags to be attached to all AWS resources"
+  type = map(string)
+}
+
+variable bucket_name {
+  description = "The logical name of the S3 bucket."
+  type = string
+}
+
+variable "deny_unencrypted_uploads" {
+  description = "Controls if uploads of unencrypted objects are denied. Default: unencrypted uploads are denied."
+  type        = bool
+}