+++ modules/database/postgresql/rds/variables.tf
@@ -65,7 +65,7 @@ variable db_storage_type {
 variable postgresql_version {
   description = "PostgreSQL version"
   type = string
-  default = "14.6"
+  default = "14.7"
 }
 
 variable vpc_id {