+++ modules/storage/blob/main.tf
@@ -15,6 +15,8 @@ locals {
   module_common_tags = merge(var.common_tags, { TerraformModuleName = "storage/blob" })
 }
 
-data aws_region current {
+data "aws_region" "current" {
   name = var.region_name
 }
+
+data "aws_caller_identity" "current" {}